+++ Stardew64Installer/Program.cs
@@ -51,74 +51,94 @@ namespace Stardew64Installer
         *********/
         private static void Prompt()
         {
-            string option = WriteReadLine("[1] I don't have a copy of the Linux version!\n[2] I'm good to go!");
-
-            if (!int.TryParse(option, out int optionNum))
-                return;
-
-            if (optionNum < 1 || optionNum > 2)
+            while (true)
             {
-                ParseFailure();
-                return;
+                string option = WriteReadLine("[1] I don't have a copy of the Linux version!\n[2] I'm good to go!").Trim();
+
+                if (option == "1")
+                {
+                    // TODO: Integrate this into this program... eventually?
+                    Console.WriteLine();
+                    Console.WriteLine("Please download DepotDownloader through https://github.com/SteamRE/DepotDownloader");
+                    WriteReadKey("Press any key to exit...");
+                    break;
+                }
+                if (option == "2")
+                {
+                    Install();
+                    break;
+                }
+
+                Console.WriteLine();
+                Console.WriteLine("We failed to parse your response. Please choose a number between 1 and 2.");
+            }
+        }
+
+        /// <summary>Interactively install to the game folder.</summary>
+        private static void Install()
+        {
+            while (true)
+            {
+                // get install path
+                Console.WriteLine();
+                string installPath = WriteReadLine("Please provide the location of the depot-downloaded copy of Stardew Valley:");
+                if (string.IsNullOrWhiteSpace(installPath))
+                    continue;
+
+                // get directory
+                DirectoryInfo installDir;
+                try
+                {
+                    installDir = new DirectoryInfo(installPath);
+                }
+                catch (Exception ex)
+                {
+                    SetColor(ConsoleColor.Red, () => Console.WriteLine($"Could not access that folder: {ex.Message}"));
+                    continue;
+                }
+                if (!installDir.Exists)
+                {
+                    SetColor(ConsoleColor.Red, () => Console.WriteLine("That folder doesn't seem to exist."));
+                    continue;
+                }
+
+                // install
+                Console.WriteLine();
+                bool installed =
+                    TryPrepareStagingFolder(installDir, out DirectoryInfo stagingDir)
+                    && TryApplyPatches(stagingDir)
+                    && TryInstallFiles(stagingDir, installDir);
+
+                if (installed)
+                    break;
             }
 
-            if (optionNum == 1)
-                DepotDownloadMessage();
-            else
-                Continue();
-        }
-
-        private static void ParseFailure()
-        {
-            Console.WriteLine();
-            Console.WriteLine("We failed to parse your response. Please choose a number between 1 and 2.");
-            Prompt();
-        }
-
-        private static void DepotDownloadMessage()
-        {
-            // TODO: Integrate this into this program... eventually?
-            Console.WriteLine();
-            Console.WriteLine("Please download DepotDownloader through https://github.com/SteamRE/DepotDownloader");
+            SetColor(ConsoleColor.Green, () => Console.WriteLine($"Installation complete! Please launch {ExeName} from the game folder."));
             WriteReadKey("Press any key to exit...");
         }
 
-        private static void Continue()
+        /// <summary>Write a message and wait for the user to hit enter.</summary>
+        /// <param name="message">The message to log.</param>
+        private static string WriteReadLine(string message)
         {
-            Console.WriteLine();
-
-            string installPath = WriteReadLine("Please provide the location of the depot-downloaded copy of Stardew Valley:");
-            Console.WriteLine();
-
-            DirectoryInfo installDir = new DirectoryInfo(installPath);
-            DirectoryInfo stagingDir = PrepareStagingFolder(installDir);
-            ApplyPatches(stagingDir);
-            InstallFiles(stagingDir, installDir);
-
-            SetColor(ConsoleColor.Green, () =>
-                Console.WriteLine($"Installation complete! Please launch {ExeName} from the game folder.")
-            );
-            WriteReadKey("Press any key to exit...");
-        }
-
-        private static string WriteReadLine(string value)
-        {
-            Console.WriteLine(value);
+            Console.WriteLine(message);
             return Console.ReadLine();
         }
 
-        private static void WriteReadKey(string value)
+        /// <summary>Write a message and wait for the user to press any key.</summary>
+        /// <param name="message">The message to log.</param>
+        private static void WriteReadKey(string message)
         {
-            Console.WriteLine(value);
+            Console.WriteLine(message);
             Console.ReadKey();
         }
 
         /// <summary>Create the staging folder with all the files needed to run the patcher.</summary>
         /// <param name="installDir">The game install folder.</param>
-        private static DirectoryInfo PrepareStagingFolder(DirectoryInfo installDir)
+        private static bool TryPrepareStagingFolder(DirectoryInfo installDir, out DirectoryInfo stagingDir)
         {
             Console.WriteLine($"Copying files to temporary folder ({StagingPath})...");
-            var stagingDir = new DirectoryInfo(StagingPath);
+            stagingDir = new DirectoryInfo(StagingPath);
 
             // copy installer files
             new DirectoryInfo(InstallerPath).RecursiveCopyTo(stagingDir.FullName);
@@ -134,7 +154,7 @@ namespace Stardew64Installer
                 {
                     Console.WriteLine($"Could not locate {ExeName}");
                     Console.WriteLine("Falling back to previous prompt...");
-                    Continue();
+                    return false;
                 }
 
                 file.CopyToAndWait(Path.Combine(stagingDir.FullName, ExeName));
@@ -145,12 +165,12 @@ namespace Stardew64Installer
                 dll.CopyToAndWait(Path.Combine(stagingDir.FullName, dll.Name));
 
             Console.WriteLine();
-            return stagingDir;
+            return true;
         }
 
         /// <summary>Patch the files in the staging folder.</summary>
         /// <param name="stagingDir">The staging folder to patch.</param>
-        private static void ApplyPatches(DirectoryInfo stagingDir)
+        private static bool TryApplyPatches(DirectoryInfo stagingDir)
         {
             // apply MonoMod patches
             Console.WriteLine($"Applying MonoMod patches to {ExeName}...");
@@ -164,12 +184,14 @@ namespace Stardew64Installer
             Console.WriteLine($"Patching {modifiedExeName} flags with CorFlags... (If this doesn't work, please relaunch with administrator privileges.)");
             RunCommand($"{Path.Combine("libs", "CorFlags.exe")} {modifiedExeName} /32BITREQ-", workingPath: stagingDir.FullName);
             Console.WriteLine();
+
+            return true;
         }
 
         /// <summary>Copy the modified files into the game folder.</summary>
         /// <param name="stagingDir">The staging folder which was patched.</param>
         /// <param name="installDir">The game install folder.</param>
-        private static void InstallFiles(DirectoryInfo stagingDir, DirectoryInfo installDir)
+        private static bool TryInstallFiles(DirectoryInfo stagingDir, DirectoryInfo installDir)
         {
             // copy override files
             Console.WriteLine("Copying override files...");
@@ -195,6 +217,7 @@ namespace Stardew64Installer
             }
 
             Console.WriteLine();
+            return true;
         }
 
         /// <summary>Run a command through <c>cmd.exe</c> and wait for it to finish.</summary>