+++ RELEASE-NOTES.md
@@ -1,6 +1,7 @@
 ﻿← [README](README.md)
 
 ## Upcoming release
+* Added support for patching the same game folder more than once (thanks to Pathoschild!).
 * Improved error-handling for game path (thanks to Pathoschild!).
 * Fixed issue where a failed install would ask for a new game path, install to that folder, then try to resume the previous failed install (thanks to Pathoschild!).